backend/requirements.txt
tensorflow==2.15.0
numpy
matplotlib
flask
flask-cors
pillow
